Support for all Intuos1/2 tools and other fixes by Chris Dragon of Dracoventions.com - Unique serial number of individual Intuos 1 and 2 tools are now passed to Wacom driver through Intuos2 emulation. - All Intuos 1 and 2 tools should now work through Intuos2 emulation.  Intuos1 4D Mouse was tested, other tools (2D Mouse, Airbrush, and Lens Cursor) were tested as much as possible without owning the actual tools. - Fix to handling of "proximity" bit from "Protocol 5" serial tablets (may not actually affect anything). - HID_listen output is easier to read and can be optionally turned off by undefining DEBUG_SERIAL_DATA and/or DEBUG_USB_DATA - printHex extended to work with 32 bit numbers at 8 hex digits.

+++ WaxB_Adb2Usb/console.h
@@ -30,7 +30,7 @@ namespace console
 	void printbit(bool bit);
 	void printNumber(long number);
 	void printBinaryByte(uint8_t number);
-	void printHex(uint16_t number, uint8_t hexdigits);
+	void printHex(uint32_t number, uint8_t hexdigits);
 	void flush();
 }
